requirements.txt
Flask>=2.0.0
flask-cors>=3.0.0
langchain>=0.0.100
langchain-core>=0.0.1
langchain-groq>=0.1.0
langchain-community>=0.0.1
langchain-google-genai>=0.0.1
langchain-cohere>=0.0.1
faiss-cpu>=1.7.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
cohere>=4.8.0
google-cloud-aiplatform>=1.32.0  # Replace with the correct service if needed
google-cloud-storage>=2.10.0    # Example for GCP storage
google-cloud-bigquery>=3.10.0   # Example for GCP BigQuery
pypdf
scikit-learn